Move test funcs that load entire config to servicetest
Updates https://github.com/open-telemetry/opentelemetry-collector/issues/4605
