+++ service/internal/builder/receivers_builder_test.go
@@ -26,10 +26,10 @@ import (
 	"go.opentelemetry.io/collector/component"
 	"go.opentelemetry.io/collector/component/componenttest"
 	"go.opentelemetry.io/collector/config"
-	"go.opentelemetry.io/collector/config/configtest"
 	"go.opentelemetry.io/collector/internal/testcomponents"
 	"go.opentelemetry.io/collector/internal/testdata"
 	"go.opentelemetry.io/collector/model/pdata"
+	"go.opentelemetry.io/collector/service/servicetest"
 )
 
 type testCase struct {
@@ -90,7 +90,7 @@ func testReceivers(t *testing.T, test testCase) {
 	factories, err := testcomponents.ExampleComponents()
 	assert.NoError(t, err)
 
-	cfg, err := configtest.LoadConfigAndValidate(path.Join("testdata", "pipelines_builder.yaml"), factories)
+	cfg, err := servicetest.LoadConfigAndValidate(path.Join("testdata", "pipelines_builder.yaml"), factories)
 	require.NoError(t, err)
 
 	// Build the pipeline
@@ -275,7 +275,7 @@ func TestBuildReceivers_Unused(t *testing.T) {
 	factories, err := testcomponents.ExampleComponents()
 	assert.NoError(t, err)
 
-	cfg, err := configtest.LoadConfigAndValidate(path.Join("testdata", "unused_receiver.yaml"), factories)
+	cfg, err := servicetest.LoadConfigAndValidate(path.Join("testdata", "unused_receiver.yaml"), factories)
 	assert.NoError(t, err)
 
 	// Build the pipeline
@@ -311,7 +311,7 @@ func TestBuildReceivers_NotSupportedDataType(t *testing.T) {
 	for _, test := range tests {
 		t.Run(test.configFile, func(t *testing.T) {
 
-			cfg, err := configtest.LoadConfigAndValidate(path.Join("testdata", test.configFile), factories)
+			cfg, err := servicetest.LoadConfigAndValidate(path.Join("testdata", test.configFile), factories)
 			assert.NoError(t, err)
 			require.NotNil(t, cfg)
